Fixed a few minor warnings on import Fixed failed message
git-svn-id: file:///Users/Shitiz/Downloads/lssvn/source/stable_plus@3653 b72ed6b6-b9f8-46b5-92b4-906544132732

+++ admin/importquestion.php
@@ -93,18 +93,18 @@ elseif
     }
 else    // unknown file - show error message
   {
-      if ($importingfrom == "http")
+      if (isset($importingfrom) && $importingfrom == "http")
       {
-          $importsurvey .= "<strong><font color='red'>".$clang->gT("Error")."</font></strong><br />\n";
-          $importsurvey .= $clang->gT("This file is not a LimeSurvey question file. Import failed.")."<br /><br />\n";
-          $importsurvey .= "</font></td></tr></table>\n";
-          $importsurvey .= "</body>\n</html>\n";
+          $importquestion .= "<strong><font color='red'>".$clang->gT("Error")."</font></strong><br />\n";
+          $importquestion .= $clang->gT("This file is not a LimeSurvey question file. Import failed.")."<br /><br />\n";
+          $importquestion .= "</font></td></tr></table>\n";
+          $importquestion .= "</body>\n</html>\n";
           unlink($the_full_file_path);
           return;
       }
       else 
       {
-          echo $clang->gT("This file is not a LimeSurvey question file. Import failed.")."\n";
+          $importquestion.= $clang->gT("This file is not a LimeSurvey question file. Import failed.")."\n";
           unlink($the_full_file_path);
           return;
       }
@@ -112,10 +112,10 @@ else    // unknown file - show error message
 
 if ($importversion != $dbversionnumber)
 {
-    $importsurvey .= "<strong><font color='red'>".$clang->gT("Error")."</font></strong><br />\n";
-    $importsurvey .= $clang->gT("Sorry, importing questions is limited to the same version. Import failed.")."<br /><br />\n";
-    $importsurvey .= "</font></td></tr></table>\n";
-    $importsurvey .= "</body>\n</html>\n";
+    $importquestion .= "<strong><font color='red'>".$clang->gT("Error")."</font></strong><br />\n";
+    $importquestion .= $clang->gT("Sorry, importing questions is limited to the same version. Import failed.")."<br /><br />\n";
+    $importquestion .= "</font></td></tr></table>\n";
+    $importquestion .= "</body>\n</html>\n";
     unlink($the_full_file_path);
     return;
 }
